ajusteComponentes

+++ src/app/views/diagnostico/diagnostico.component.ts
@@ -4,8 +4,12 @@ import { Router } from '@angular/router';
 import { DiagnosticoCrudComponent } from 'src/app/components/diagnostico-crud/diagnostico-crud.component';
 import { HeaderService } from 'src/app/components/template/header/header.service';
 import { Diagnostico } from 'src/app/models/diagnostico.model';
+import { Mascota } from 'src/app/models/mascota.model';
+import { Veterinario } from 'src/app/models/veterinario.model';
 import { DiagnosticoService } from 'src/app/services/diagnostico.service';
 import { MascotaService } from 'src/app/services/mascota.service';
+import { VeterinarioService } from 'src/app/services/veterinario.service';
+import { take } from 'rxjs';
 
 @Component({
   selector: 'app-diagnostico',
@@ -14,23 +18,21 @@ import { MascotaService } from 'src/app/services/mascota.service';
 })
 export class DiagnosticoComponent implements OnInit {
   diagnosticos: Diagnostico[];
-  displayedColumns = [
-    'id',
-    'fecha',
-    'estado',
-    'enfermedad',
-    'tratamiento',
-    'veterinario',
-    'action',
-  ];
+  displayedColumns = ['fecha', 'estado', 'enfermedad', 'tratamiento', 'action'];
   nuevoDiagnostico: Diagnostico;
   editDiagnostico: Diagnostico;
   modal: string;
   idDiagnostico: string;
-  mascotaDiagnosticos: any;
+  mascotaDiagnosticos: Diagnostico[];
+  mascota: Mascota;
   idMascota: any;
+  idHidden: boolean;
+  comboVeterinario: Veterinario[];
+  loading: boolean;
+  veterinarioDiagnostico: Veterinario[];
 
   constructor(
+    private veterinarioService: VeterinarioService,
     private mascotaService: MascotaService,
     private diagnosticoService: DiagnosticoService,
     private router: Router,
@@ -38,7 +40,7 @@ export class DiagnosticoComponent implements OnInit {
     public dialog: MatDialog
   ) {
     headerService.headerData = {
-      title: 'Listado de diagnosticos',
+      title: 'Historial de mascota',
       icon: 'storefront',
       routeUrl: '/diagnostico',
     };
@@ -46,12 +48,18 @@ export class DiagnosticoComponent implements OnInit {
     this.modal = '';
     this.editDiagnostico = {};
     this.idDiagnostico = '';
-
+    this.loading = false;
     this.nuevoDiagnostico = {};
     this.diagnosticos = [];
+    this.mascotaDiagnosticos = [];
+    this.idHidden = true;
+    this.comboVeterinario = [];
+    this.mascota = {};
+    this.veterinarioDiagnostico = [];
   }
 
   ngOnInit(): void {
+    this.listVeterinario();
     this.cargarList();
   }
 
@@ -62,16 +70,25 @@ export class DiagnosticoComponent implements OnInit {
     this.modal = '';
 
     const dialogo1 = this.dialog.open(DiagnosticoCrudComponent, {
-      data: (this.nuevoDiagnostico = {
-
-      }),
+      data: (this.nuevoDiagnostico = {}),
     });
 
     dialogo1.afterClosed().subscribe((diag) => {
-      this.diagnosticoService.create(diag).subscribe(() => {
-        this.diagnosticoService.showMessage('Creado con éxito!');
-        this.router.navigate(['/diagnostico']);
-      });
+      this.loading = false;
+      try {
+        if (diag) {
+          diag.mascota = this.mascota;
+          diag.mascota.id = Number(this.idMascota);
+
+          this.diagnosticoService.create(diag).subscribe(() => {
+            this.diagnosticoService.showMessage('Creado con éxito!');
+            this.router.navigate(['/diagnostico']);
+            this.cargarList();
+          });
+        } else {
+          this.cargarList();
+        }
+      } catch (error) {}
     });
   }
 
@@ -80,17 +97,27 @@ export class DiagnosticoComponent implements OnInit {
     sessionStorage.removeItem('modal');
     sessionStorage.setItem('modal', this.modal);
     this.modal = '';
-    this.editDiagnostico = diag;
+    // this.diagnosticoService.getVetporIdDiagnostico(diag.id).subscribe((v) => {
+    //   diag.veterinario = v;
+    // });
 
+    this.editDiagnostico = diag;
     const dialogo1 = this.dialog.open(DiagnosticoCrudComponent, {
       data: this.editDiagnostico,
     });
 
     dialogo1.afterClosed().subscribe((diag) => {
-      this.diagnosticoService.update(diag).subscribe(() => {
-        this.diagnosticoService.showMessage('Datos actualizados!');
-        this.router.navigate(['/diagnostico']);
-      });
+      try {
+        if (diag) {
+          this.diagnosticoService.update(diag).subscribe(() => {
+            this.diagnosticoService.showMessage('Datos actualizados!');
+            this.router.navigate(['/diagnostico']);
+            this.cargarList();
+          });
+        } else {
+          this.cargarList();
+        }
+      } catch (error) {}
     });
   }
 
@@ -99,6 +126,7 @@ export class DiagnosticoComponent implements OnInit {
     sessionStorage.removeItem('modal');
     sessionStorage.setItem('modal', this.modal);
     this.modal = '';
+
     this.editDiagnostico = diag;
 
     const dialogo1 = this.dialog.open(DiagnosticoCrudComponent, {
@@ -118,19 +146,28 @@ export class DiagnosticoComponent implements OnInit {
     });
 
     dialogo1.afterClosed().subscribe((diag) => {
-      this.diagnosticoService.delete(diag.id).subscribe(() => {
-        this.diagnosticoService.showMessage('Eliminado con exito');
-        this.router.navigate(['/diagnostico']);
-      });
+      this.loading = false;
+
+      if (diag) {
+        this.diagnosticoService.delete(diag.id).subscribe(() => {
+          this.diagnosticoService.showMessage('Eliminado con exito');
+          this.router.navigate(['/diagnostico']);
+          this.cargarList();
+        });
+      } else {
+        this.cargarList();
+      }
     });
   }
 
-  agregar(diag: Diagnostico) {
-    this.diagnosticos.push(this.nuevoDiagnostico);
-  }
-
-  update(diag: Diagnostico) {
-    this.diagnosticos.push(this.nuevoDiagnostico);
+  listVeterinario() {
+    this.veterinarioService
+      .getAllVet()
+      .pipe(take(1))
+      .subscribe((data) => {
+        this.comboVeterinario = data;
+        // this.loading = false
+      });
   }
 
   cargarList() {
@@ -138,10 +175,18 @@ export class DiagnosticoComponent implements OnInit {
 
     this.mascotaService.diagnosticoById(this.idMascota).subscribe((md) => {
       let aux = md;
-console.log(aux)
       this.mascotaDiagnosticos = aux;
-    });
-  }
+      this.mascotaDiagnosticos.forEach((element) => {
+        let aux = element['fecha'];
+        let anio = aux?.substring(0, 4);
+        let mes = aux?.substring(5, 7);
+        let dia = aux?.substring(8, 10);
+
+        element['fecha'] = `${dia}/${mes}/${anio}`;
+      });
+    }); // servicio
+    this.loading = false;
+  } // funcion carga
 
   navigateToMascotaCrud(diag: Diagnostico): void {
     // this.idDiagnostico = (diag.id).toString();
